Refactor of code into smaller modules
