--- src/cleanslate.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-rm -f ../*.blockchain
