+++ src/node.ts
@@ -1,0 +1,13 @@
+export class Node {
+  public id: string;
+  public url: URL;
+
+  constructor(id: string, url: URL) {
+    this.id = id;
+    this.url = url;
+  }
+
+  public toString(): string {
+    return `${this.id}:${this.url}`;
+  }
+}
