+++ src/shell-scripts/cleanslate.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+rm -f ../../*.blockchain
