+++ src/workshop/05_hash_function.ts
