+++ src/workshop/06_pow.ts
