+++ src/workshop/11_mining_api.ts
